+++ cmd/local/apicollect/jobprofiles.go
@@ -24,9 +24,9 @@ import (
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/conf"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/queriesjson"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/restclient"
+	"github.com/dremio/dremio-diagnostic-collector/cmd/local/threading"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
 	"github.com/dremio/dremio-diagnostic-collector/pkg/ddcio"
-	"github.com/dremio/dremio-diagnostic-collector/pkg/threading"
 )
 
 func getNumberOfJobProfilesCollected(c *conf.CollectConf) (tried, collected int, err error) {
