+++ cmd/local/local.go
@@ -40,9 +40,9 @@ import (
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/nodeinfocollect"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
 
+	"github.com/dremio/dremio-diagnostic-collector/cmd/local/threading"
 	"github.com/dremio/dremio-diagnostic-collector/pkg/ddcio"
 	"github.com/dremio/dremio-diagnostic-collector/pkg/masking"
-	"github.com/dremio/dremio-diagnostic-collector/pkg/threading"
 	"github.com/dremio/dremio-diagnostic-collector/pkg/versions"
 )
